+++ src/realm/sync/config.cpp
@@ -87,6 +87,8 @@ SimplifiedProtocolError get_simplified_error(sync::ProtocolError err)
         // Connection level errors
         case ProtocolError::connection_closed:
         case ProtocolError::other_error:
+        case ProtocolError::switch_to_flx_sync:
+        case ProtocolError::switch_to_pbs:
             // Not real errors, don't need to be reported to the binding.
             return SimplifiedProtocolError::ConnectionIssue;
         case ProtocolError::unknown_message:
@@ -103,7 +105,6 @@ SimplifiedProtocolError get_simplified_error(sync::ProtocolError err)
         case ProtocolError::bad_changeset:
         case ProtocolError::bad_changeset_header_syntax:
         case ProtocolError::bad_changeset_size:
-        case ProtocolError::bad_changesets:
         case ProtocolError::bad_decompression:
         case ProtocolError::unsupported_session_feature:
         case ProtocolError::transact_before_upload: