+++ src/realm/sync/noinst/client_impl_base.hpp
@@ -249,7 +249,8 @@ private:
                                            bool verify_servers_ssl_certificate,
                                            util::Optional<std::string> ssl_trust_certificate_path,
                                            std::function<SyncConfig::SSLVerifyCallback>,
-                                           util::Optional<SyncConfig::ProxyConfig>, bool& was_created);
+                                           util::Optional<SyncConfig::ProxyConfig>, SyncServerMode,
+                                           bool& was_created);
 
     // Destroys the specified connection.
     void remove_connection(ClientImpl::Connection&) noexcept;
@@ -294,6 +295,7 @@ enum class ClientImpl::ConnectionTerminationReason {
     server_said_try_again_later,       ///< Client received ERROR message with try_again=yes
     server_said_do_not_reconnect,      ///< Client received ERROR message with try_again=no
     pong_timeout,                      ///< Client did not receive PONG after PING
+    switch_wire_protocols,             ///< Client should reconnect with a different wire protocols
 
     /// The application requested a feature that is unavailable in the
     /// negotiated protocol version.
@@ -401,7 +403,7 @@ public:
     Connection(ClientImpl&, connection_ident_type, ServerEndpoint, const std::string& authorization_header_name,
                const std::map<std::string, std::string>& custom_http_headers, bool verify_servers_ssl_certificate,
                util::Optional<std::string> ssl_trust_certificate_path, std::function<SSLVerifyCallback>,
-               util::Optional<ProxyConfig>, ReconnectInfo);
+               util::Optional<ProxyConfig>, ReconnectInfo, SyncServerMode);
 
     ~Connection();
 
@@ -466,11 +468,12 @@ private:
     void change_state_to_disconnected() noexcept;
 
     // These are only called from ClientProtocol class.
+    bool is_flx_sync_connection() const noexcept;
     void receive_pong(milliseconds_type timestamp);
     void receive_error_message(int error_code, StringData message, bool try_again, session_ident_type);
     void receive_ident_message(session_ident_type, SaltedFileIdent);
     void receive_download_message(session_ident_type, const SyncProgress&, std::uint_fast64_t downloadable_bytes,
-                                  const ReceivedChangesets&);
+                                  int64_t query_version, DownloadBatchState batch_state, const ReceivedChangesets&);
     void receive_mark_message(session_ident_type, request_ident_type);
     void receive_unbound_message(session_ident_type);
     void handle_protocol_error(ClientProtocol::Error);
@@ -504,6 +507,8 @@ private:
     const util::Optional<ProxyConfig> m_proxy_config;
     ReconnectInfo m_reconnect_info;
     int m_negotiated_protocol_version = 0;
+    SyncServerMode m_sync_mode = SyncServerMode::PBS;
+    bool m_is_flx_sync_connection = false;
 
     ConnectionState m_state = ConnectionState::disconnected;
 
@@ -760,7 +765,8 @@ public:
     /// event loop thread of the associated client object, the specified history
     /// accessor must **not** be the one made available by access_realm().
     bool integrate_changesets(ClientReplication&, const SyncProgress&, std::uint_fast64_t downloadable_bytes,
-                              const ReceivedChangesets&, VersionInfo&, IntegrationError&);
+                              const ReceivedChangesets&, VersionInfo&, IntegrationError&,
+                              DownloadBatchState last_in_batch);
 
     /// To be used in connection with implementations of
     /// initiate_integrate_changesets().
@@ -773,7 +779,7 @@ public:
     /// (Connection::activate_session()), or after initiation of deactivation
     /// (Connection::initiate_session_deactivation()).
     void on_changesets_integrated(bool success, version_type client_version, DownloadCursor download_progress,
-                                  IntegrationError error);
+                                  IntegrationError error, DownloadBatchState batch_state);
 
     void on_connection_state_changed(ConnectionState, const SessionErrorInfo*);
 
@@ -878,7 +884,8 @@ private:
     ///
     /// This function is guaranteed to not be called before activation, and also
     /// not after initiation of deactivation.
-    void initiate_integrate_changesets(std::uint_fast64_t downloadable_bytes, const ReceivedChangesets&);
+    void initiate_integrate_changesets(std::uint_fast64_t downloadable_bytes, DownloadBatchState batch_state,
+                                       const ReceivedChangesets&);
 
     /// See request_upload_completion_notification().
     ///
@@ -927,6 +934,8 @@ private:
     // uploaded to the server. Set to true when uploading of the token has been
     // initiated via a BIND or a REFRESH message.
     bool m_access_token_sent = false;
+
+    DownloadBatchState m_download_batch_state = DownloadBatchState::LastInBatch;
 
     // Set to true when download completion is reached. Set to false after a
     // slow reconnect, such that the upload process will become suspended until
@@ -1082,7 +1091,7 @@ private:
     void send_unbind_message();
     std::error_code receive_ident_message(SaltedFileIdent);
     void receive_download_message(const SyncProgress&, std::uint_fast64_t downloadable_bytes,
-                                  const ReceivedChangesets&);
+                                  DownloadBatchState last_in_batch, int64_t query_version, const ReceivedChangesets&);
     std::error_code receive_mark_message(request_ident_type);
     std::error_code receive_unbound_message();
     std::error_code receive_error_message(int error_code, StringData message, bool try_again);
@@ -1274,6 +1283,7 @@ inline bool ClientImpl::Connection::was_voluntary(ConnectionTerminationReason re
         case ConnectionTerminationReason::pong_timeout:
         case ConnectionTerminationReason::missing_protocol_feature:
         case ConnectionTerminationReason::http_tunnel_failed:
+        case ConnectionTerminationReason::switch_wire_protocols:
             break;
     }
     return false;
@@ -1480,6 +1490,7 @@ inline void ClientImpl::Session::reset_protocol_state() noexcept
     m_upload_progress = m_progress.upload;
     m_last_version_selected_for_upload = m_upload_progress.client_version;
     m_last_download_mark_sent          = m_last_download_mark_received;
+    m_download_batch_state = DownloadBatchState::LastInBatch;
     // clang-format on
 }