+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### Internals
 * The 'power' unary operator template to be used in a query expression is removed
 * Renamed ClientResyncMode::SeamlessLoss -> DiscardLocal.
+* Updated sync client to be able to open connections to FLX sync-enabled apps in baas ([#5009](https://github.com/realm/realm-core/pull/5009))
 
 ----------------------------------------------